feat: enhance security and user management with helmet integration and request validation
- Integrated helmet middleware in the main application for improved security against common web vulnerabilities.
- Updated user, job, and category controllers to utilize authenticated request types, ensuring user context is available in command executions.
- Refactored command and query structures to include requesting user IDs, enhancing data integrity and access control.
- Modified request DTOs for job and category creation and updates to enforce validation rules, ensuring data consistency.
- Improved error handling in command handlers to prevent unauthorized access to resources.

+++ src/application/commands/handlers/category/deleteCategory.handler.ts
@@ -2,11 +2,12 @@ import { Inject, NotFoundException } from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { type ICategoryRepository } from 'src/domain/interfaces/ICategoryRepository';
 import { CategoryRepositorySymbol } from 'src/modules/symbols/symbols';
-import { ApiErrorMessages } from 'src/shared/constants/api-error-messages';
 import { DeleteCategoryCommand } from '../../category.commands';
 
 @CommandHandler(DeleteCategoryCommand)
-export class DeleteCategoryHandler implements ICommandHandler<DeleteCategoryCommand> {
+export class DeleteCategoryHandler
+  implements ICommandHandler<DeleteCategoryCommand>
+{
   constructor(
     @Inject(CategoryRepositorySymbol)
     private readonly _category_repository: ICategoryRepository,
@@ -14,10 +15,8 @@ export class DeleteCategoryHandler implements ICommandHandler<DeleteCategoryComm
 
   async execute(command: DeleteCategoryCommand): Promise<boolean> {
     const category = await this._category_repository.findById(command.id);
-    if (!category) {
-      throw new NotFoundException(
-        ApiErrorMessages.category.notFoundForId(command.id),
-      );
+    if (!category || category.user.id !== command.requestingUserId) {
+      throw new NotFoundException();
     }
     return this._category_repository.remove(command.id);
   }
